Proofread kdoc and rename `test` package to `testing`
